+++ tiny-lib/MerkleTree.hpp
@@ -27,7 +27,7 @@ public:
 
 private:
 	static std::vector<std::vector<std::shared_ptr<MerkleNode>>> Chunk(
-		const std::vector<std::shared_ptr<MerkleNode>>& nodes, size_t chunkSize);
+		const std::vector<std::shared_ptr<MerkleNode>>& nodes, uint32_t chunkSize);
 
 	static std::shared_ptr<MerkleNode> FindRooot(const std::vector<std::shared_ptr<MerkleNode>>& nodes);
 };
